Removed the remaining size_t references that affected the FIFE API and replaced them with more appropriate definitions.  The remaining size_t definitions will remain as they wont cause any issues.  fixes[t:152]

+++ engine/core/model/structures/map.cpp
@@ -78,7 +78,7 @@ namespace FIFE {
 		throw NotFound(id);
 	}
 
-	size_t Map::getNumLayers() const {
+	uint32_t Map::getNumLayers() const {
 		return m_layers.size();
 	}
